analyze_agent/requirements.txt
pyyaml>=6.0
nats-py>=2.7.0
regex>=2023.5.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-ollama>=0.1.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
aiohttp>=3.9.0
asyncio-mqtt>=0.13.0
pydantic>=2.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
numpy>=1.21.0
transformers>=4.21.0
# tf-keras  # 如果系统安装了Keras 3，需要安装此包以保证兼容性